+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 - Optimization Cat Editor options panel
 - Fix Timeline list scroll gesture conflicts [#654](https://github.com/hushenghao/AndroidEasterEggs/issues/654)
+- Upgrade the minimum supported version to Android 6.0
+- Update translations
+- Upgrade project dependencies
 
 ### v4.3.1 (2025-12-01)